changed some stuff regarding the overall css of login
