--- frontend/src/components/LoginSignUp/LoginSignUp.css
@@ -1,156 +1,0 @@
-.container {
-    position: absolute;
-    left: 33%;
-    display: flex;
-    flex-direction: column;
-    width: 600px;
-    background: #78b336;
-    padding-bottom: 30px;
-    border-radius: 12px;
-}
-
-
-.header{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 9px;
-    width: 100%;
-    margin-top: 30px;
-}
-
-.text{
-    color: #3A4A9C;
-    font-size: 48px;
-    font-weight: 700;
-}
-
-.underline{
-    width: 61px;
-    height: 6px;
-    background: #3A4A9C;
-    border-radius: 9px;
-}
-
-.inputs{
-    margin-top: 55px;
-    display: flex;
-    flex-direction: column;
-    gap: 25px;
-}
-
-.input{
-    display: flex;
-    align-items: center;
-    margin: auto;
-    width: 480px;
-    height: 80px;
-    background: #eaeaea;
-    border-radius: 6px;
-}
-
-.input img{
-    margin: 0px 30px;
-}
-
-.input input{
-    height: 50px;
-    width: 400px;
-    background: transparent;
-    border: none;
-    outline: none;
-    color: #797979;
-    font-size: 19px;
-}
-
-.forgot-password{
-    padding-left: 62px;
-    margin-top: 27px;
-    color: #797979;
-    font-size: 18px;
-}
-
-.forgot-password span{
-    color: #3A4A9C;
-    cursor: pointer;
-}
-
-.submit-container{
-    display: flex;
-    gap: 30px;
-    margin: 60px auto;
-}
-
-.submit{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 220px;
-    height: 59px;
-    color: #fff;
-    background: #3A4A9C;
-    border-radius: 50px;
-    font-size: 19px;
-    font-weight: 700;
-    cursor: pointer;
-}
-
-.gray{
-    background: #EAEAEA;
-    color: #676767;
-}
-.login-wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    color: #78b336;
-}
-
-.submitButton {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 480px;
-    height: 59px;
-    color: #fff;
-    background: #3A4A9C;
-    border-radius: 50px;
-    font-size: 19px;
-    font-weight: 700;
-    cursor: pointer;
-    margin: 1px auto 0 auto;
-}
-.toast {
-    position: fixed;
-    top: 10%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    min-width: 200px;
-    max-width: 80%;
-    padding: 16px 24px;
-    border-radius: 8px;
-    text-align: center;
-    font-weight: 500;
-    z-index: 9999;
-    opacity: 0;
-    animation: fadeInOut 3s ease forwards;
-    pointer-events: none;
-}
-.toast.success {
-    color: #1a472a;
-    background: #e6ffed;
-    border: 1px solid #b7f5c5;
-}
-
-.toast.error {
-    color: #611a15;
-    background: #fdecea;
-    border: 1px solid #f5c2c0;
-}
-@keyframes fadeInOut {
-    0%   { opacity: 0; }
-    10%  { opacity: 1; }
-    90%  { opacity: 1; }
-    100% { opacity: 0; }
-}
