+++ frontend/src/style/login-signupStyle.css
@@ -1,0 +1,121 @@
+.login-wrapper {
+    height: 100%;
+    width: 100%;
+    background: transparent;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-family: 'Segoe UI', sans-serif;
+    box-sizing: border-box;
+    overflow: hidden;
+}
+
+.container {
+    background-color: #ffffff;
+    border-radius: 12px;
+    padding: 40px 35px;
+    box-shadow: 0 8px 25px rgba(0, 0, 0, 0.1);
+    width: 400px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.header .text {
+    font-size: 28px;
+    font-weight: bold;
+    color: #2c3e50;
+    text-align: center;
+}
+
+.underline {
+    width: 50px;
+    height: 4px;
+    background-color: #2c3e50;
+    border-radius: 2px;
+    margin: 10px auto 20px auto;
+}
+
+.inputs {
+    width: 100%;
+}
+
+.input {
+    display: flex;
+    align-items: center;
+    background-color: #f4f4f4;
+    border-radius: 8px;
+    padding: 10px;
+    margin-bottom: 15px;
+}
+
+.input img {
+    width: 20px;
+    margin-right: 10px;
+}
+
+.input input {
+    border: none;
+    background: none;
+    outline: none;
+    flex: 1;
+    font-size: 16px;
+    color: #2c3e50;
+}
+
+.input input::placeholder {
+    color: #888;
+}
+
+.submit-container {
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    margin-top: 20px;
+}
+
+.submit {
+    flex: 1;
+    background-color: #4CAF50;
+    color: white;
+    padding: 12px;
+    text-align: center;
+    border-radius: 6px;
+    cursor: pointer;
+    margin: 0 5px;
+    transition: background-color 0.3s ease;
+}
+
+.submit:hover {
+    background-color: #43a047;
+}
+
+.toggle-mode {
+    margin-top: 20px;
+    font-size: 14px;
+    text-align: center;
+}
+
+.toggle-mode span {
+    color: #2c3e50;
+    cursor: pointer;
+    text-decoration: underline;
+}
+
+.toast {
+    margin-top: 15px;
+    text-align: center;
+    padding: 10px;
+    border-radius: 6px;
+    font-weight: bold;
+}
+
+.toast.success {
+    background-color: #c8e6c9;
+    color: #2e7d32;
+}
+
+.toast.error {
+    background-color: #ffcdd2;
+    color: #c62828;
+}
